fix(web): move the reset-widths button into the Appearance menu
lups2000 and emanuele-em felt it didn't belong in the Flow List tab
since it's unrelated to search/highlight/intercept; the Options
menu's existing Appearance group is a better home.
