+++ web/src/js/__tests__/components/Header/MainMenuSpec.tsx
@@ -1,29 +1,8 @@
 import * as React from "react";
 import FlowListMenu from "../../../components/Header/FlowListMenu";
-import { fireEvent, render, screen } from "../../test-utils";
-import { TStore, testState } from "../../ducks/tutils";
-
-const withColumnWidths = () =>
-    TStore({
-        ...testState,
-        ui: { ...testState.ui, columnWidths: { size: 70 } },
-    });
+import { render } from "../../test-utils";
 
 test("MainMenu", () => {
     const { asFragment } = render(<FlowListMenu />);
     expect(asFragment()).toMatchSnapshot();
-});
-
-test("MainMenu resets the flow table column widths", () => {
-    const { store } = render(<FlowListMenu />, { store: withColumnWidths() });
-
-    fireEvent.click(screen.getByText("Reset Widths"));
-
-    expect(store.getState().ui.columnWidths).toEqual({});
-});
-
-test("MainMenu has nothing to reset until a column is resized", () => {
-    render(<FlowListMenu />);
-
-    expect(screen.getByText("Reset Widths").closest("button")).toBeDisabled();
 });
